Create agent.py
The agent is the heart part of the project.  it should be run separately before other parts to create the files (CSV, TXT,...) which visuals.py uses to create the charts. two log folders need to be created one in this directory and the other one in the parent directory. The agent.py creates the files in the log folder in this directory. These files should be copied to the log folder of the parent directory before running visuals.py